+updated: the preloaded SQLite DB --> +updated: Cart table to use REAL in price field, instead of INTEGER to accommodate decimal values +updated: in UsbongDbHelper, updateCartTable(...) --> The app is now able to successfully insert all product items into the internal SQLite DB's Cart table. --> TODO: -put: in UsbongUtils, the portion of code that creates from the raw list, two lists: 1) unique product items, 2) the quantity of each unique product item --> This is because I use it more than once; the other is in CartActivity. --> The index used for the product list, i.e. tempList (to mean "temporary list"), matches with the index for the quantity list. --> This is to properly identify the product item in both lists.

+++ assets/usbong_store.sql
@@ -167,7 +167,7 @@ CREATE TABLE cart (
   cart_id INTEGER PRIMARY KEY,
   product_id INTEGER,
   quantity INTEGER,
-  price INTEGER,
+  price REAL,
   purchased_datetime_stamp TEXT
 );
